+++ src/components/Search/Search.module.css
@@ -1,0 +1,51 @@
+.searchPage{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  width: 100%;
+}
+.search {
+  width: 40%;
+  margin: 4rem 0 2rem;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  padding: 0.3rem 1.5rem;
+  border-radius: 3rem;
+  background-image: linear-gradient(to right, #434343 0%, black 100%);
+}
+.search input {
+  flex: 1;
+  border: none;
+  font-size: 1.5rem;
+  font-family: var(--font-raleway);
+  font-weight: 500;
+  padding-right: 1rem;
+  outline: none;
+  color: #e0e0e0;
+  background: inherit;
+  padding: 1rem;
+}
+
+.search img {
+  width: 35px;
+  height: 35px;
+  cursor: pointer;
+}
+.search img:hover{
+  transition: all 0.2s linear;  
+  transform: scale(1.1, 1.1);
+}
+.search:focus-within{
+  box-shadow: 0px 0px 28px  rgba(55, 228, 133, 0.719);
+}
+
+.container {
+  display: flex; 
+  justify-content: space-around;
+  flex-wrap: wrap;
+
+}
